Fix color-scheme installation

+++ tasks/vim.rake
@@ -1,16 +1,18 @@
 namespace :vim do
+  HOME = ENV['HOME']
+  COLORSCHEMES = {
+    "clouds-midnight" => "http://darcsden.com/alex/dotfiles/raw/.vim/colors/clouds-midnight.vim"
+  }
+
   def install_colorscheme(cs=nil)
     if cs
-      colorschemes = {
-        "clouds-midnight" => "http://darcsden.com/alex/dotfiles/raw/.vim/colors/clouds-midnight.vim"
-      }
-      colorscheme = colorschemes[cs]
+      colorscheme = COLORSCHEMES[cs]
       if colorscheme
-        FileUtils.cd("#{ENV['HOME']}/.vim/bundle/vim-colorschemes/colors") do
+        FileUtils.cd(File.join(HOME, ".vim/bundle/vim-colorschemes/colors")) do
           sh("wget #{colorscheme}")
         end
+        sh("echo colorscheme #{cs} >> ~/.vimrc")
       end
-      sh("echo colorscheme #{cs} >> ~/.vimrc")
     end
   end
 
@@ -22,13 +24,11 @@ namespace :vim do
 
   desc "Install vim"
   task :vim_spf13 => ["vim"] do
-    #pkgs = %w{vim}
-    #install_pkg(pkgs)
-    FileUtils.cd("#{ENV['HOME']}") do
-      sh("curl http://j.mp/spf13-vim3 -L -o - | sh")
+    FileUtils.cd(HOME) do
+      sh("curl http://j.mp/spf13-vim3 -L -o - | sh") unless File.exists?(".vimrc-bundles")
 
       # Install clouds-mignight colorscheme
-      install_colorscheme('cloud-midnight')
+      install_colorscheme('clouds-midnight')
 
       notice("Installed vim spf13")
     end
@@ -37,4 +37,3 @@ end
 
 task :vim => ["vim:vim", "vim:vim_spf13"] # , "vim:vimtools"]
 
-
